requirements.txt
openai
pandas
google-api-python-client
google-auth
google-auth-oauthlib
gitpython
requests
python-dotenv